+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:fitsSystemWindows="true"
-    tools:context="com.valevich.moneytracker.MainActivity">
+    tools:context=".ui.activities.MainActivity">
 
     <include layout="@layout/content_main"/>
