+++ app/src/main/res/layout/fragment_categories.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="com.valevich.moneytracker.fragments.CategoriesFragment">
+    tools:context=".ui.fragments.CategoriesFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/categories_list"
