+++ app/src/main/res/layout/fragment_settings.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="com.valevich.moneytracker.fragments.SettingsFragment"
+             tools:context=".ui.fragments.SettingsFragment"
              android:background="@android:color/holo_red_light">
