+++ app/src/main/res/layout/fragment_statistics.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="com.valevich.moneytracker.fragments.StatisticsFragment"
+             tools:context=".ui.fragments.StatisticsFragment"
              android:background="@android:color/holo_orange_light">
 
 </FrameLayout>
